move Owned{Up,}Value to new file
